requirements.txt
langchain==0.1.13
streamlit==1.32.2
chromadb==0.4.24
pypdf==4.1.0

# FastAPI stack
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9

# Embeddings
sentence-transformers>=2.5
langchain-huggingface>=0.0.9
huggingface_hub>=0.24
